requirements.txt
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
asyncio==4.0.0
bleak==2.1.1
bottle==0.12.25
cachetools==4.2.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==2.0.12
cryptography==46.0.4
DateTime==6.0
dbus-fast==3.1.2
gigachat==0.1.43
googleapis-common-protos==1.72.0
grpcio==1.71.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.83
langchain-gigachat==0.3.12
langsmith==0.6.8
logging==0.4.9.6
orjson==3.11.7
packaging==25.0
protobuf==5.29.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
python-dotenv==0.19.0
python-telegram-bot==22.6
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
RPi.GPIO==0.7.1
schedule==1.1.0
setuptools==80.10.2
six==1.17.0
tenacity==9.1.3
tornado==6.5.4
types-requests==2.32.4.20260107
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==3.0
urllib3==2.6.3
uuid_utils==0.14.0
xxhash==3.6.0
yandexcloud==0.295.0
zope.interface==8.2
zstandard==0.25.0
